django_backend/requirements.txt
djangorestframework
pymongo
python-dotenv
geopy
django
PyJWT
django-cors-headers
PyPDF2
PyMuPDF
python-docx
boto3
langchain
langchain_groq
langchain-google-genai
langchain-community
langchain-core
faiss-cpu
pandas
openpyxl
pinecone-client
bleach
docx2python
pillow
fpdf
paramiko
django_extensions
usaddress
googlemaps